Ball class cleanup
I had it so the x and y speeds were recalculating each frame, which was unnecessary. Now the ball's x and y speed only recalculates each time it hits a wall. This calculation is done under "calculateComponentSpeeds" method. This method will also be used when the ball hits a paddle.